move buzzer-related code into its own file
Add app_buzzer.c/h and move buzzer-related code there. This file will
always be built, but functions will only be included for original Thingy:91
(the thingy91 has a buzzer, thingy91x does not).
