+++ src/app_sensors.c
@@ -14,14 +14,9 @@ LOG_MODULE_REGISTER(app_sensors, LOG_LEVEL_DBG);
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/drivers/sensor.h>
 #include <zephyr/device.h>
-#include <zephyr/drivers/pwm.h>
 
 #include "app_sensors.h"
 #include "app_settings.h"
-
-#define FUNKYTOWN_NOTES 13
-#define MARIO_NOTES	37
-#define GOLIOTH_NOTES	21
 
 static struct golioth_client *client;
 
@@ -31,200 +26,6 @@ static const struct gpio_dt_spec user_led = GPIO_DT_SPEC_GET(DT_ALIAS(user_led),
 const struct device *light = DEVICE_DT_GET_ONE(rohm_bh1749);
 const struct device *weather = DEVICE_DT_GET_ONE(bosch_bme680);
 const struct device *accel = DEVICE_DT_GET_ONE(adi_adxl362);
-
-const struct pwm_dt_spec sBuzzer = PWM_DT_SPEC_GET(DT_ALIAS(buzzer_pwm));
-
-enum song_choice {
-	beep,
-	funkytown,
-	mario,
-	golioth
-};
-
-enum song_choice song = 3;
-
-struct note_duration {
-	int note;     /* hz */
-	int duration; /* msec */
-};
-
-struct note_duration funkytown_song[FUNKYTOWN_NOTES] = {
-	{.note = C5, .duration = quarter},
-	{.note = REST, .duration = eigth},
-	{.note = C5, .duration = quarter},
-	{.note = Bb4, .duration = quarter},
-	{.note = C5, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = G4, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = G4, .duration = quarter},
-	{.note = C5, .duration = quarter},
-	{.note = F5, .duration = quarter},
-	{.note = E5, .duration = quarter},
-	{.note = C5, .duration = quarter}};
-
-struct note_duration mario_song[MARIO_NOTES] = {
-	{.note = E6, .duration = quarter},
-	{.note = REST, .duration = eigth},
-	{.note = E6, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = E6, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = C6, .duration = quarter},
-	{.note = E6, .duration = half},
-	{.note = G6, .duration = half},
-	{.note = REST, .duration = quarter},
-	{.note = G4, .duration = whole},
-	{.note = REST, .duration = whole},
-	/* break in sound */
-	{.note = C6, .duration = half},
-	{.note = REST, .duration = quarter},
-	{.note = G5, .duration = half},
-	{.note = REST, .duration = quarter},
-	{.note = E5, .duration = half},
-	{.note = REST, .duration = quarter},
-	{.note = A5, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = B5, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = Bb5, .duration = quarter},
-	{.note = A5, .duration = half},
-	{.note = G5, .duration = quarter},
-	{.note = E6, .duration = quarter},
-	{.note = G6, .duration = quarter},
-	{.note = A6, .duration = half},
-	{.note = F6, .duration = quarter},
-	{.note = G6, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = E6, .duration = quarter},
-	{.note = REST, .duration = quarter},
-	{.note = C6, .duration = quarter},
-	{.note = D6, .duration = quarter},
-	{.note = B5, .duration = quarter}};
-
-struct note_duration golioth_song[] = {
-	{.note = C6, .duration = quarter},
-	{.note = REST, .duration = 100},
-	{.note = G5, .duration = 100},
-	{.note = A5, .duration = 100},
-	{.note = Bb5, .duration = 100},
-	{.note = REST, .duration = 100},
-	{.note = Bb5, .duration = 100},
-	{.note = REST, .duration = quarter},
-	{.note = C5, .duration = half},
-	{.note = REST, .duration = half},
-	{.note = REST, .duration = quarter},
-	{.note = C6, .duration = quarter}
-};
-
-/* Thread plays song on buzzer */
-
-K_SEM_DEFINE(buzzer_initialized_sem, 0, 1); /* Wait until buzzer is ready */
-
-#define BUZZER_STACK 1024
-
-extern void buzzer_thread(void *d0, void *d1, void *d2)
-{
-	/* Block until buzzer is available */
-	k_sem_take(&buzzer_initialized_sem, K_FOREVER);
-	while (1) {
-		switch (song) {
-		case 0:
-			LOG_DBG("beep");
-			pwm_set_dt(&sBuzzer, PWM_HZ(1000), PWM_HZ(1000) / 2);
-			k_msleep(100);
-			break;
-		case 1:
-			LOG_DBG("funkytown");
-			for (int i = 0; i < FUNKYTOWN_NOTES; i++) {
-				if (funkytown_song[i].note < 10) {
-					/* Low frequency notes represent a 'pause' */
-					pwm_set_pulse_dt(&sBuzzer, 0);
-					k_msleep(funkytown_song[i].duration);
-				} else {
-					pwm_set_dt(&sBuzzer, PWM_HZ(funkytown_song[i].note),
-						   PWM_HZ((funkytown_song[i].note)) / 2);
-					k_msleep(funkytown_song[i].duration);
-				}
-			}
-			break;
-
-		case 2:
-			LOG_DBG("mario");
-			for (int i = 0; i < MARIO_NOTES; i++) {
-				if (mario_song[i].note < 10) {
-					/* Low frequency notes represent a 'pause' */
-					pwm_set_pulse_dt(&sBuzzer, 0);
-					k_msleep(mario_song[i].duration);
-				} else {
-					pwm_set_dt(&sBuzzer, PWM_HZ(mario_song[i].note),
-						   PWM_HZ((mario_song[i].note)) / 2);
-					k_msleep(mario_song[i].duration);
-				}
-			}
-			break;
-		case 3:
-			LOG_DBG("golioth");
-			for (int i = 0; i < (sizeof(golioth_song) / sizeof(golioth_song[1])); i++) {
-				if (golioth_song[i].note < 10) {
-					/* Low frequency notes represent a 'pause' */
-					pwm_set_pulse_dt(&sBuzzer, 0);
-					k_msleep(golioth_song[i].duration);
-				} else {
-					pwm_set_dt(&sBuzzer, PWM_HZ(golioth_song[i].note),
-						   PWM_HZ((golioth_song[i].note)) / 2);
-					k_msleep(golioth_song[i].duration);
-				}
-			}
-			break;
-		default:
-			LOG_WRN("invalid switch state");
-			break;
-		}
-
-		/* turn buzzer off (pulse duty to 0) */
-		pwm_set_pulse_dt(&sBuzzer, 0);
-
-		/* Sleep thread until awoken externally */
-		k_sleep(K_FOREVER);
-	}
-}
-
-int app_buzzer_init(void)
-{
-	if (!device_is_ready(sBuzzer.dev)) {
-		return -ENODEV;
-	}
-	k_sem_give(&buzzer_initialized_sem);
-	return 0;
-}
-
-K_THREAD_DEFINE(buzzer_tid, BUZZER_STACK, buzzer_thread, NULL, NULL, NULL,
-		K_LOWEST_APPLICATION_THREAD_PRIO, 0, 0);
-
-void play_beep_once(void)
-{
-	song = beep;
-	k_wakeup(buzzer_tid);
-}
-
-void play_funkytown_once(void)
-{
-	song = funkytown;
-	k_wakeup(buzzer_tid);
-}
-
-void play_mario_once(void)
-{
-	song = mario;
-	k_wakeup(buzzer_tid);
-}
-
-void play_golioth_once(void)
-{
-	song = golioth;
-	k_wakeup(buzzer_tid);
-}
 
 /* Set (unset) LED indicators for user action */
 void user_led_set(uint8_t state)
