+++ src/app_sensors.h
@@ -12,58 +12,4 @@
 void app_sensors_set_client(struct golioth_client *sensors_client);
 void app_sensors_read_and_stream(void);
 
-int app_buzzer_init(void);
-void play_funkytown_once(void);
-void play_mario_once(void);
-void play_beep_once(void);
-void play_golioth_once(void);
-
-#define BUZZER_MAX_FREQ 2500
-#define BUZZER_MIN_FREQ 75
-
-#define sixteenth 38
-#define eigth	  75
-#define quarter	  150
-#define half	  300
-#define whole	  600
-
-#define C4  262
-#define Db4 277
-#define D4  294
-#define Eb4 311
-#define E4  330
-#define F4  349
-#define Gb4 370
-#define G4  392
-#define Ab4 415
-#define A4  440
-#define Bb4 466
-#define B4  494
-#define C5  523
-#define Db5 554
-#define D5  587
-#define Eb5 622
-#define E5  659
-#define F5  698
-#define Gb5 740
-#define G5  784
-#define Ab5 831
-#define A5  880
-#define Bb5 932
-#define B5  988
-#define C6  1046
-#define Db6 1109
-#define D6  1175
-#define Eb6 1245
-#define E6  1319
-#define F6  1397
-#define Gb6 1480
-#define G6  1568
-#define Ab6 1661
-#define A6  1760
-#define Bb6 1865
-#define B6  1976
-
-#define REST 1
-
 #endif /* __APP_SENSORS_H__ */
